Introduced BaseSetting root Java class and extracted Setting interface to better implement the delegator pattern
