+++ logstash-core/src/main/java/org/logstash/settings/Coercible.java
@@ -21,7 +21,7 @@ package org.logstash.settings;
 
 import java.util.function.Predicate;
 
-public abstract class Coercible<T> extends Setting<T> {
+public abstract class Coercible<T> extends BaseSetting<T> {
     public Coercible(String name, T defaultValue, boolean strict, Predicate<T> validator) {
         super(name, strict, validator);
